Add message originator addresses to peerstore
Put the message originator's addresses into the pubsub message.  When this message is received by go-legs, the addresses are added to the peerstore.  This allows datatransfer to make a connection back to the provider to get advertisements and content.  This is necessary when the provider and indexer are not directly connected, but are connected through intermediate nodes.