Update head on explicit `Sync` only if no CID nor selector is specified (#40)
When the explicit `Sync` function is called only update the current head
if no CID nor selector is given. In this case, `Sync` acts similarly to
triggering a on-off explicit sync cycle for the latest head; it:
1. Fetches the latest head from the remote publisher,
2. Uses the default selector sequence wrapped with a logic to stop at
   currently synced head, and
3. Updates the current head to the latest.

If either CID or selector are specified then the current head is not
updated.

The rationale for this change is to allow legs to be used for ad-hoc
syncing of partial DAGs that would have otherwise been skipped by the
default selector sequence used by the background sync loop.

Consistently reflect the above approach on both HTTP and libp2p
endpoints.

Add tests that assert the internal reference to the current head is
indeed updated if `Sync` is called without CID nor selector.