Do not continue with explicit sync when there is no root CID
`Subscriber.Sync` queries the latest root CID if one is not given. Check
that the result of query is an actual CID before proceeding with the
sync. If there is no latest root CID, i.e. query for the latest CID
returns `cid.Undef`, then return `cid.Undef` with no error.

Fixes #64