Do not log error and return same error (#77)
* Do not log error and return same error

Logging an error and then returning that error leads to logging the same errors multiple times.  This decreases the density of useful information and decreases the value of logs.  By returning an error and logging it at a higher level, the error message can contain more meaningful information about the error.

- Log consistently using structured logging
- Update datatransfer and libp2p