Add optional miner ID to pubsub message (#80)
* Add optional extra data to pubsub message

The extra data can be used to carry whatever is needed by the target recipient(s).  For indexing, this will be to carry a miner ID.  By carrying the miner ID in the pubsub message the lotus chain-nodes that relay the messages can lookup the merket process peer ID and ensure that was the peer that signed the message.  This authenticates the message as being sent from a peer that is known on-chain.