Add CheckAlreadySynced sync option
This is necessary to check if already synced to the current head, when a custom selector is passed in.  If the selector has a stop node, the selector only compares the previous link against the stop node.  So, if the current head is the stop node, this needs a special case check.  Normally, when a custom selector is passed in this special case check is not done (since legs does not know what the selector does).  This option lets the caller tell legs to do the special case check.