Fix possible deadlock if user places backpressure
If a user placed backpressure by not pulling on the OnSyncFinished event
channel then we would prevent any handler from running for that
publisher.

This could lead to a deadlock like the one in storetheindex. Where the
indexer applies backpressure if the ingest workers cannot keep up. The
ingest workers were deadlocked because they could not finish their work
because it required another call to `subscriber.Sync` to fetch the ad
entries. But the sync was blocked on a handler that was blocked on
sending to a channel that was blocked by the indexer pulling the channel
that was blocked by not having a worker available to send work to that
was blocked by the worker not being to make progress...