Fix rate limit burst refresh for graphsync
After dt/gs session hits the burst limit, it is necessary wait for the token bucket to refil before restarting the session. The wait was happening but was also emptying its bucket.  So when the session restarted there was no burst and rate limiting would happen immidiately after reading the next block or two.  This fixes this to wait without emptying the bucket.

Other changes:
- Improved logging
- AllowPeerFunc only retruns bool. It is up to the function to log any errors, not legs.  Also returning the error does not allow any decision beyond what the boolean return value already does.