Fix rate limit burst refresh for graphsync (#124)
* Fix rate limit burst refresh for graphsync

After dt/gs session hits the burst limit, it is necessary to wait for the token bucket to refill before restarting the session. Previously, the wait was happening but was also emptying its bucket.  So when the session restarted, the token bucket was empty and rate-limiting happened immediately when reading the next block.  This fixes this to wait without emptying the bucket.

Other changes:
- Improved logging
- AllowPeerFunc only returns bool. It is up to the function to log any errors, not legs.  Also returning the error does not allow any decision beyond what the boolean return value already does.