Implement the ability to sync a DAG in segments of configurable depth
Syncing very large DAGs in a single sync call, either via `httpsync` or
`dtsync` can result in excessive memory usage doe to internal state
stored by underlying sync mechanism. For example, `httpsync` keeps the
traversed path in memory, which can result in high memory usage when the
syncing excessively large dags.

To avoid such issues, break up the syncing of a DAG into a series of
individual syncs with a smaller and configurable depth limit. This would
then provide an opportunity for any ephemeral state stored in memory
within each sync to be garbage collected, reducing the overall memory
footprint when syncing large DAGs.

The changes introduced are:
* An option to control the maximum depth limit for each sync segment.
* an extension to the block hook that lets the user control the flow of
  segmented sync cycles.
* scoped block hook overrides the general block hook if both are set.

For backward compatibility, segmented sync is disabled by default. The
segmented sync is only enabled if:
* the segment depth limit is not a negative value, and
* a block hook, either general or scoped, is set.

The block hook is needed because in go-legs level, we do not have enough
information to decode a synced block and determine what its "next" link
may be. Therefore, the user of the library is required to do this and
control the flow via the segment block hook actions.

The scoped block hook overrides the general block hook, because, the
"actions" extension to the block hook modifies the state of sync.
Running both general and scoped block hook could then make conflicting
calls to the action and result in unexpected behaviour. To avoid such
edge cases, the scoped hook simply overrides the general hook if both
are set. The tests are adjusted to assert this behaviour.

Relates to: https://github.com/filecoin-project/storetheindex/issues/487