Set the rate limiter resolver function for subscriber only
The rate limiter resolver function returns a rate-limiter for a given peerID. This is called at the start of every sync. Since this function can return a different rate limiter for any sync, there is no need for an individual sync to override this function. Removing this override capability removes unneeded complexity.