Set rate limiter on syncer only
Do not set rate limiter for dssync.Sync or httpsync.Sync. Instead only set this when the Syncer for the specific sync is created.

Return to allowing a separate rate limiter to be set for an individual sync. This may be useful when a peer has a particular rete limiter, but a single sync needs to use a different rate limiter.  While this could be done within the rateLimiterFor function passed to the Subscriber, it may be much simpler for the application to simply pass a different rate limiter when calling Subscriber.Sync. With other changes here, allowing this does not add any significant complexity.