Simplify rate limiter configuration (#130)
* Set rate limiter to use for sync only when creating Syncer

Do not set rate limiter for `dssync.Sync` or `httpsync.Sync`. Instead only set this when creating the `Syncer` for the specific sync. Use the rate-limiter provided for that individual sync, or if one was not provided, get it from the Subscriber's `rateLimiterFor` function if that function is available. If no rate-limiter is specified, then use a nil rate-limiter, which avoids any rate-limiting logic.

* Specifying a nil rate limiter is now OK

It is not necessary to construct an infinite rate limiter when there is no rate-limiting. A nil rate limiter avoids any rate-limiting logic.

* Stabilize tests by closing libp2p host connections

Tests were failing on some systems due to an error "socket: too many open files". This was caused by libp2p hosts having connections that were not closed. Tests have been modified to close the connections.

* Add comments and normalize comment formatting