Remove handlers after they have been idle for some time
When publishers change or go out of service, the go-legs handlers created to hold state for those publishers need to be removed since they will not be used again. This PR removed idle handlers to prevent accumulating handlers for defunct publishers.

There is also a new `RemoveHandlers` API that can be used to immediately remove a handler when it is known that the handler is no longer needed.