Handle already registered dtsync vouchers gracefully
It is possible for the same host to be a legs publisher and a lengs
subscriber. Instantiating both on the same host with an existing
datatransfer manager will attempt to register the vouchers twice which
results in a failure.

Match the body of the error returned by datatransfer and handle it
gracefully. Note that currently that seems to be the best we can do
since there are no APIs exposed by the manager to check if a type is
already registered or re-register a voucher type.

Fixes: https://github.com/filecoin-project/go-legs/issues/144