Verify integrity of blocks synced by httpsync prior to persisting
Require that the hash of blocks downloaded by httpsync match the digest
of their CID, with one exception: be lenient in accepting digest of
`DagJSON` codec; in that, accept both the digest calculated from the
spec-compliant serialization of the content, as well as one calculated
from directly ingesting the entire response body.

Include tests that assert leniency in both the cases using sample data
downloaded from NFT.Storage and the legs HTTP publisher which is indeed
spec-compliant.

See:
 - https://ipld.io/specs/codecs/dag-json/spec/#serialization

Fixes #159