Call blockhook for CIDs that may have been found locally
Add a mechanism to signal to the blockhook the CIDs that were found
locally without exchange over graphsync channel. This is to make sure
the contract of blockhook is obeyed: it is called for all CIDs
encountered during a sync, regardless of where they were found.

This assures that mechanisms reliant on blockhook will continue to
function even after the optimisation introduced in #115.

Add tests that assert the fix works for both fully and partially present
DAGs.

Fixes #161